+++ src/main/java/com/kap/mechanics_api/dto/movimentacaoestoque/RegistroSaidaMovimentacaoEstoqueRequestDTO.java
@@ -12,6 +12,6 @@ public record RegistroSaidaMovimentacaoEstoqueRequestDTO(
         Integer quantidade,
 
         @NotNull(message = "Informe a ordem de servico")
-        Long ordemServicoId
+        Integer ordemServicoId
 ) {
 }
