+++ src/main/java/com/kap/mechanics_api/exception/MovimentacaoEstoqueNaoEncontradaException.java
@@ -2,7 +2,7 @@ package com.kap.mechanics_api.exception;
 
 public class MovimentacaoEstoqueNaoEncontradaException extends RuntimeException {
 
-    public MovimentacaoEstoqueNaoEncontradaException(Long id) {
+    public MovimentacaoEstoqueNaoEncontradaException(Integer id) {
         super("Movimentacao de estoque nao encontrada com o id " + id);
     }
 }
