+++ src/main/java/com/kap/mechanics_api/exception/OrdemServicoNaoEncontradaException.java
@@ -2,7 +2,7 @@ package com.kap.mechanics_api.exception;
 
 public class OrdemServicoNaoEncontradaException extends RuntimeException {
 
-    public OrdemServicoNaoEncontradaException(Long id) {
+    public OrdemServicoNaoEncontradaException(Integer id) {
         super("Ordem de servico nao encontrada com o id " + id);
     }
 }
