+++ src/main/java/com/kap/mechanics_api/exception/OrdemServicoNaoEstaEmExecucaoException.java
@@ -2,7 +2,7 @@ package com.kap.mechanics_api.exception;
 
 public class OrdemServicoNaoEstaEmExecucaoException extends RuntimeException {
 
-    public OrdemServicoNaoEstaEmExecucaoException(Long ordemServicoId) {
+    public OrdemServicoNaoEstaEmExecucaoException(Integer ordemServicoId) {
         super("A ordem de servico " + ordemServicoId + " nao esta em execucao.");
     }
 }
