+++ src/main/java/com/kap/mechanics_api/repository/MovimentacaoEstoqueRepository.java
@@ -7,13 +7,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface MovimentacaoEstoqueRepository extends JpaRepository<MovimentacaoEstoque, Long> {
+public interface MovimentacaoEstoqueRepository extends JpaRepository<MovimentacaoEstoque, Integer> {
 
     List<MovimentacaoEstoque> findAllByOrderByDataHoraDesc();
 
     List<MovimentacaoEstoque> findByItemEstoque_IdOrderByDataHoraDesc(Integer itemEstoqueId);
 
-    List<MovimentacaoEstoque> findByOrdemServico_IdOrderByDataHoraDesc(Long ordemServicoId);
+    List<MovimentacaoEstoque> findByOrdemServico_IdOrderByDataHoraDesc(Integer ordemServicoId);
 
     List<MovimentacaoEstoque> findByTipoOrderByDataHoraDesc(TipoMovimentacaoEstoque tipo);
